perf: store hits in a redis store
Instead of storing all requests in-memory, we store them in a redis store. This will help improve performance, since there will be more free memory on the servers. The previous approach was also very flawed, since it does not share state across servers nor processes. The redis store solves that as well