fix(Channels): flatten all channels when some are updated
Previously, channels could only be moved up and down within their category, so the mendChannelPositions method only cared about the channels in a specific category. The feature that allowed users to move channels between categories also introduced a bug where positions would not be flattened if a channel was moved from one category to another. This commit fixes that, since the mendChannelPositions method now mends the positions of the channels in all categories, including the categories themselves.