perf(getPresentableObject): move query population to pre find hook
This heavily reduces the amount of database queries required to populate the fields. The original method, when fetching i.e. a channel, would result in about 3 database queries *per channel*. If we fetched about 100 channels at the same time, which is not uncommon, we would have to perform about 300 queries just for population. This perfomance update will reduce it to around 3 extra queries *in total*.